Adding a results DIV
I also moved the script tag down to the end of the HTML document which
is considered good practice. It allows the page to fully load before
having the user download scripts, so the user isn't waiting with a blank
screen. Additionally I have changed the input text tag value attribute
to the new HTML5 placeholder attribute. It is more user-friendly, not
supported in older browsers, so we may need to look into doing a
graceful fall-back using modernizer and jquery.